Schedule API optimisations
Move some methods from CLI to API to help with schedule editing
